apps/core/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlmodel
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
firebase-admin
requests
httpx
python-dotenv
